Fix runtime compilation of vue components
Some views were partially broken as we were missing the vue bundler library.
This impacted:
 - Admin Stacks
 - Admin Templates
 - Team User Invites
 - User Invites
 - Project Billing

This PR moves the inline components into their own component files. A couple
were copy/paste clones, so they now share the same common component.

I have not modified the billing ones - left them as inline.
