+++ frontend/src/pages/admin/Users/Invitations.vue
@@ -17,20 +17,7 @@ import { markRaw } from 'vue'
 import FormHeading from '@/components/FormHeading'
 import ItemTable from '@/components/tables/ItemTable'
 import InviteUserCell from '@/components/tables/cells/InviteUserCell'
-import { TrashIcon } from '@heroicons/vue/outline'
-
-const MemberInviteRemoveButton = {
-    template: '<ff-button @click="removeInvite"><template v-slot="icon-left"><TrashIcon/></template></ff-button>',
-    props: ['id', 'team', 'onremove'],
-    components: {
-        TrashIcon
-    },
-    methods: {
-        async removeInvite () {
-            await this.onremove(this.team.id, this.id)
-        }
-    }
-}
+import MemberInviteRemoveButton from './components/MemberInviteRemoveButton'
 
 export default {
     name: 'UserInviteTable',
