+++ frontend/src/pages/team/Projects.vue
@@ -26,14 +26,9 @@ import { markRaw } from 'vue'
 import { Roles } from '@core/lib/roles'
 import teamApi from '@/api/team'
 import ItemTable from '@/components/tables/ItemTable'
-import { ExternalLinkIcon, PlusSmIcon } from '@heroicons/vue/outline'
+import { PlusSmIcon } from '@heroicons/vue/outline'
 
 import ProjectStatusBadge from '@/pages/project/components/ProjectStatusBadge'
-
-const OpenEditorLink = {
-    template: '<ExternalLinkIcon class="w-4 h-4"/>',
-    components: { ExternalLinkIcon }
-}
 
 export default {
     name: 'TeamProjects',
@@ -43,8 +38,7 @@ export default {
             columns: [
                 { name: 'Name', class: ['flex-grow'], property: 'name', link: 'link' },
                 { name: 'Status', class: ['w-44'], component: { is: markRaw(ProjectStatusBadge) } },
-                { name: 'Updated', class: ['w-44', 'text-xs'], property: 'updatedSince' },
-                { name: '', class: ['w-14'], linkClass: ['forge-button-inline', 'px-2', 'py-2'], component: { is: markRaw(OpenEditorLink) }, external: true, link: 'url' }
+                { name: 'Updated', class: ['w-44', 'text-xs'], property: 'updatedSince' }
             ]
         }
     },
